pyproject.toml
[project]
name = "pyHost"
version = "0.0.1"
dependencies = [
    "numpy",
    "pycryptodomex",
    "crcmod",
    "pyserial",
    "pyside6",
]
[project.scripts]
pyhost = "pyHost:main"
pyhost-cli = "pyHost:main_cli"